Move solid shell testing to own test file
