+++ tests/reference-files/test_integration_four_c_solid_shell_thickness_direction_block.4C.yaml
