+++ tests/test_four_c.py
@@ -37,10 +37,6 @@ from beamme.four_c.element_beam import Beam3rHerm2Line3
 from beamme.four_c.input_file import InputFile
 from beamme.four_c.locsys_condition import LocSysCondition
 from beamme.four_c.model_importer import import_four_c_model
-from beamme.four_c.solid_shell_thickness_direction import (
-    get_visualization_third_parameter_direction_hex8,
-    set_solid_shell_thickness_direction,
-)
 from beamme.mesh_creation_functions.beam_helix import create_beam_mesh_helix
 from beamme.mesh_creation_functions.beam_line import create_beam_mesh_line
 from beamme.utils.nodes import is_node_on_plane
@@ -258,97 +254,6 @@ def test_four_c_beam_potential_helix(
     input_file.add(mesh)
 
     assert_results_close(get_corresponding_reference_file_path(), input_file)
-
-
-def test_four_c_solid_shell_direction_detection(
-    get_default_test_beam_material,
-    assert_results_close,
-    get_corresponding_reference_file_path,
-    tmp_path,
-):
-    """Test the solid shell direction detection functionality."""
-
-    # Test the plates
-    _, mesh_block = import_four_c_model(
-        input_file_path=get_corresponding_reference_file_path(
-            reference_file_base_name="test_other_create_cubit_input_files_solid_shell_blocks"
-        ),
-        convert_input_to_mesh=True,
-    )
-
-    # Add a beam element to check the function also works with beam elements
-    create_beam_mesh_line(
-        mesh_block,
-        Beam3rHerm2Line3,
-        get_default_test_beam_material(material_type="reissner"),
-        [0, 0, 0],
-        [1, 0, 0],
-        n_el=1,
-    )
-    # Set the thickness direction and compare result
-    set_solid_shell_thickness_direction(mesh_block.elements, selection_type="thickness")
-    assert_results_close(
-        get_corresponding_reference_file_path(additional_identifier="blocks"),
-        mesh_block,
-    )
-
-    # Test the dome
-    _, mesh_dome_original = import_four_c_model(
-        input_file_path=get_corresponding_reference_file_path(
-            reference_file_base_name="test_other_create_cubit_input_files_solid_shell_dome"
-        ),
-        convert_input_to_mesh=True,
-    )
-
-    # Test that the thickness version works
-    mesh_dome = mesh_dome_original.copy()
-    set_solid_shell_thickness_direction(mesh_dome.elements, selection_type="thickness")
-    assert_results_close(
-        get_corresponding_reference_file_path(additional_identifier="dome_thickness"),
-        mesh_dome,
-    )
-
-    # Test that the direction function version works
-    def director_function(cell_center):
-        """Return director that will be used to determine the solid thickness
-        direction."""
-        return cell_center / np.linalg.norm(cell_center)
-
-    mesh_dome = mesh_dome_original.copy()
-    set_solid_shell_thickness_direction(
-        mesh_dome.elements,
-        selection_type="projection_director_function",
-        director_function=director_function,
-    )
-    assert_results_close(
-        get_corresponding_reference_file_path(additional_identifier="dome_thickness"),
-        mesh_dome,
-    )
-
-    # Test that the constant direction version works
-    mesh_dome = mesh_dome_original.copy()
-    set_solid_shell_thickness_direction(
-        mesh_dome.elements,
-        selection_type="projection_director",
-        director=[0, 0, 1],
-        identify_threshold=None,
-    )
-    assert_results_close(
-        get_corresponding_reference_file_path(
-            additional_identifier="dome_constant_direction"
-        ),
-        mesh_dome,
-    )
-
-    # Also test the visualization function
-    ref_file = get_corresponding_reference_file_path(
-        additional_identifier="dome_constant_direction", extension="vtu"
-    )
-    test_file = tmp_path / ref_file.name
-
-    grid = get_visualization_third_parameter_direction_hex8(mesh_dome)
-    grid.save(test_file)
-    assert_results_close(ref_file, test_file)
 
 
 def test_four_c_locsys_condition(
